Improves progress bars.
This makes the two progress bars match roughly what is done in Lightning, with an implementation inspired by Lightning's.

The basic idea is that the primary bar is the one making the pass over the data during the e-step; a secondary bar is generated immediately when we compute loss during the (so-called) validation step; then the loss from that validation is reported as a "postfix" of the primary bar.